requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.1
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
Flask==1.1.1
Flask-CDN==1.5.3
Flask-Cors==3.0.7
fuzzywuzzy==0.17.0
gunicorn==19.9.0
idna==2.8
imageio==2.6.1
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.3.0
nltk==3.4.5
numpy==1.16.5
opencv-python==4.1.0.25
packaging==20.4
pandas==0.25.1
pdf2image==1.6.0
pdfminer.six==20200104
pdfplumber==0.5.22
pdftotext==2.1.4
pika==1.1.0
Pillow==8.0.1
pip-check-reqs==2.1.1
pluggy==0.13.1
py==1.8.1
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.9
pyparsing==2.4.7
PyPDF4==1.27.0
pytesseract==0.2.9
pytest==5.0.1
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2020.1
PyYAML==5.1.2
requests==2.22.0
ruamel.base==1.0.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.15.0
sortedcontainers==2.3.0
tqdm==4.46.1
unicodecsv==0.14.1
urllib3==1.25.9
Wand==0.6.4
wcwidth==0.2.3
Werkzeug==0.15.5
zipp==3.1.0
